temp_deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
sqlalchemy==2.0.23
databases==0.8.0 